refractor: Create premade border and filler items
